+++ controllers/add.js
@@ -1,0 +1,12 @@
+const {Contact} = require('../models/contacts')
+
+const add = async (req, res, next) => {
+  try {
+    const result = await Contact.create(req.body);
+    res.status(201).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = add;
