--- controllers/contactsFunk.js
@@ -1,88 +1,0 @@
-const { Contact } = require("../models/contacts");
-const { HttpError } = require("../helpers/HttpError");
-
-const getAll = async (req, res, next) => {
-  try {
-    const result = await Contact.find();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getById = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    console.log(req.params.contactId);
-    console.log(req.params.id);
-    const result = await Contact.findById(contactId);
-
-    if (!result) {
-      const error = new Error("Not found");
-      error.status = 404;
-      throw error;
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const add = async (req, res, next) => {
-  try {
-    const result = await Contact.create(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateById = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateStatusContact = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-      new: true,
-    });
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const deleteById = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndDelete(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-};
-
-module.exports = {
-  getAll,
-  getById,
-  add,
-  updateById,
-  updateStatusContact,
-  deleteById,
-};
