+++ controllers/deleteById.js
@@ -1,0 +1,17 @@
+const { Contact } = require("../models/contacts");
+const {HttpError}= require('../helpers/HttpError')
+
+const deleteById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await Contact.findByIdAndDelete(contactId);
+    if (!result) {
+      throw HttpError(404, "Not found");
+    }
+    res.json({ message: "contact deleted" });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = deleteById;
