+++ controllers/getAll.js
@@ -1,0 +1,12 @@
+const {Contact} = require('../models/contacts')
+
+const getAll = async (req, res, next) => {
+  try {
+    const result = await Contact.find();
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getAll;
