+++ routes/api/contacts.js
@@ -1,9 +1,17 @@
 const express = require("express");
-const controllers = require("../../controllers/contactsFunk");
+const {
+  getAll,
+  getById,
+  add,
+  updateById,
+  updateStatusContact,
+  deleteById,
+} = require("../../controllers");
+
 const {
   validatePutBody,
-    validatePostBody,
-  validateFavorite
+  validatePostBody,
+  validateFavorite,
 } = require("../../middlewares/validateBody");
 const isValidId = require("../../middlewares/isValidId");
 
@@ -11,26 +19,26 @@ const router = express.Router();
 
 const { schemas } = require("../../models/contacts");
 
-router.get("/", controllers.getAll);
+router.get("/", getAll);
 
-router.get("/:contactId", isValidId, controllers.getById);
+router.get("/:contactId", isValidId, getById);
 
-router.post("/", validatePostBody(schemas.addSchema), controllers.add);
+router.post("/", validatePostBody(schemas.addSchema), add);
 
 router.put(
   "/:contactId",
   isValidId,
   validatePutBody(schemas.addSchema),
-  controllers.updateById
+  updateById
 );
 
 router.patch(
   "/:contactId/favorite",
   isValidId,
   validateFavorite(schemas.updateFavoriteSchema),
-  controllers.updateStatusContact
+  updateStatusContact
 );
 
-router.delete("/:contactId", isValidId, controllers.deleteById);
+router.delete("/:contactId", isValidId, deleteById);
 
 module.exports = router;
